% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psesvbody.R
\name{psesvbody}
\alias{psesvbody}
\title{Shiny dashboard body for PSESV}
\usage{
psesvbody(appvar)
}
\arguments{
\item{appvar}{An object returned from fluidPage()}
}
\description{
Includes tags and css for body o app and uses fluidRow() in dashboard body.
}
