% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psesvdashboard.R
\name{psesvdashboard}
\alias{psesvdashboard}
\title{Dashboardpage for PSESV}
\usage{
psesvdashboard(articleno, vizno, app)
}
\arguments{
\item{articleno}{Article number pulled from metadata file}

\item{vizno}{Visualization number pulled from metadata file}

\item{app}{An object returned from fluidPage()}
}
\description{
Pulls together PSESV metadata, header, sidebar and body.
}
