% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psesvmetadata.R
\name{psesvmetadata}
\alias{psesvmetadata}
\title{Metadata for PSESV}
\usage{
psesvmetadata(articleno, vizno)
}
\arguments{
\item{articleno}{Article number pulled from metadata file}

\item{vizno}{Visualization number pulled from metadata file}
}
\description{
Read in metadata from for given article number and viz number
}
