% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psesvprintmetadata.R
\name{psesvprintmetadata}
\alias{psesvprintmetadata}
\title{print metadata for sidebar}
\usage{
psesvprintmetadata(articleno, vizno)
}
\arguments{
\item{articleno}{Article number pulled from metadata file}

\item{vizno}{Visualization number pulled from metadata file}
}
\description{
print metadata for sidebar
}
