% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psesvrelatedviz.R
\name{psesvrelatedviz}
\alias{psesvrelatedviz}
\title{Print html for related PSESV visualizations}
\usage{
psesvrelatedviz(articleno, vizno)
}
\arguments{
\item{articleno}{Article number pulled from metadata file}

\item{vizno}{Visualization number pulled from metadata file}
}
\description{
generates HTML for related visualizations if they exist (other visualizations with same article number)
}
