% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psesvsiderbar.R
\name{psesvsidebar}
\alias{psesvsidebar}
\title{Shiny dashboard sidebar for PSESV}
\usage{
psesvsidebar(articleno, vizno)
}
\arguments{
\item{articleno}{Article number pulled from metadata file}

\item{vizno}{Visualization number pulled from metadata file}
}
\description{
Sets up sidebar for PSESV with article number and viz number.
}
