% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psesvvars.R
\docType{data}
\name{psesvroot}
\alias{psesvroot}
\title{PSESV root directory}
\format{An object of class \code{character} of length 1.}
\usage{
psesvroot
}
\description{
PSESV root directory
}
\keyword{datasets}
